+++ src/event/bootp.rs
@@ -286,6 +286,7 @@ impl BlocklistBootp {
 }
 
 impl Match for BlocklistBootp {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }