+++ src/event/dcerpc.rs
@@ -424,6 +424,7 @@ impl BlocklistDceRpc {
 }
 
 impl Match for BlocklistDceRpc {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }