+++ src/event/dhcp.rs
@@ -385,6 +385,7 @@ impl BlocklistDhcp {
 }
 
 impl Match for BlocklistDhcp {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }