+++ src/event/dns.rs
@@ -302,6 +302,7 @@ impl DnsCovertChannel {
 }
 
 impl Match for DnsCovertChannel {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }
@@ -487,6 +488,7 @@ impl LockyRansomware {
 }
 
 impl Match for LockyRansomware {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }
@@ -817,6 +819,7 @@ impl CryptocurrencyMiningPool {
 }
 
 impl Match for CryptocurrencyMiningPool {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }
@@ -1140,6 +1143,7 @@ impl BlocklistDns {
 }
 
 impl Match for BlocklistDns {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }