+++ src/event/kerberos.rs
@@ -275,6 +275,7 @@ impl BlocklistKerberos {
 }
 
 impl Match for BlocklistKerberos {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }