+++ src/event/malformed_dns.rs
@@ -293,6 +293,7 @@ impl BlocklistMalformedDns {
 }
 
 impl Match for BlocklistMalformedDns {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }