+++ src/event/mqtt.rs
@@ -256,6 +256,7 @@ impl BlocklistMqtt {
 }
 
 impl Match for BlocklistMqtt {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }