+++ src/event/network.rs
@@ -258,6 +258,7 @@ impl fmt::Display for NetworkThreat {
 }
 
 impl Match for NetworkThreat {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }