+++ src/event/ntlm.rs
@@ -237,6 +237,7 @@ impl BlocklistNtlm {
 }
 
 impl Match for BlocklistNtlm {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }