+++ src/event/radius.rs
@@ -311,6 +311,7 @@ impl BlocklistRadius {
 }
 
 impl Match for BlocklistRadius {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }