+++ src/event/smb.rs
@@ -285,6 +285,7 @@ impl BlocklistSmb {
 }
 
 impl Match for BlocklistSmb {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }