+++ src/event/smtp.rs
@@ -254,6 +254,7 @@ impl BlocklistSmtp {
 }
 
 impl Match for BlocklistSmtp {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }