+++ src/event/ssh.rs
@@ -304,6 +304,7 @@ impl BlocklistSsh {
 }
 
 impl Match for BlocklistSsh {
+    crate::event::common::impl_match_pair_country_codes!();
     fn src_addrs(&self) -> &[IpAddr] {
         std::slice::from_ref(&self.orig_addr)
     }